Replace expensive css animations with requestAnimationFrame.
The scrolling issue was caused by css fade in animations continuing to
run even after the animation was complete. This kept that GPU in a heavy
processing state that caused the scrolling to flicker.

The css animation fades that were causing the problem have been replaced
with a Fade component that will use requestAnimationFrame in conjuction
with componentDidMount (with logic to stop when opacity >= 1) and
componentWillUnmount (cancelling animation when unmounted) to properly
stop and start the fading of its children.